# Description

L1 jax tests occasionally timeout due to long running attention and softmax tests. This PR addresses the softmax tests.

Before:
```
================================================================================
TEST RUNTIME SUMMARY (grouped by function)
test_distributed_softmax.py
================================================================================
test_softmax                                                 | 288x | 2082.61s | avg:   7.23s
test_softmax_shardy                                          |  24x |   10.02s | avg:   0.42s
================================================================================
TOTAL RUNTIME                                                |      | 2092.63s |
================================================================================
```

After:
```
================================================================================
TEST RUNTIME SUMMARY (grouped by function)
================================================================================
test_softmax                                                 | 288x |  216.69s | avg:   0.75s
test_softmax_gspmd                                           |  24x |    9.87s | avg:   0.41s
================================================================================
TOTAL RUNTIME                                                |      |  226.56s |
================================================================================
```

## Type of change

- [ ] Documentation change (change only to the documentation, either a fix or a new content)
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [x] Infra/Build change
- [ ] Code refactoring

## Changes

- Update `tests/jax/test_distributed_softmax.py` to reduce the large shape of some inputs. This dim the softmax is being performed in is still the same size, I've only updated the dimensions that are batched over. The same parallelism is still enabled.
- Update `tests/jax/test_distributed_softmax.py` to change our main tests to use shardy and a smaller subset for GSPMD now that shardy is the default (previously on main this was reversed so our shardy test suite was smaller).
- Add a utility to `conftest.py` to print test runtimes per function by specifying `NVTE_JAX_TEST_TIMING=1` in the environment

# Checklist:

- [x] I have read and followed the [contributing guidelines](https://github.com/NVIDIA/TransformerEngine/blob/main/CONTRIBUTING.rst)
- [x] The functionality is complete
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] New and existing unit tests pass locally with my changes